Rename the parsers to `ValidationTargetParser` and `TypeNameParser`
